+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-24 17:58:43
-// Commit: fix(i18n): prevent double tap duplicate action in i18n
+// Updated at: 2026-05-24 19:57:45
+// Commit: feat(i18n): add async i18n stream provider
 
 class i18nComponentState {
-  final String commitRef = '1409';
-  final String lastMessage = 'fix(i18n): prevent double tap duplicate action in i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-24T17:58:43+05:00');
+  final String commitRef = '1416';
+  final String lastMessage = 'feat(i18n): add async i18n stream provider';
+  final DateTime timestamp = DateTime.parse('2026-05-24T19:57:45+05:00');
 }